tidy(ui): organize & document builder dnd logic
